fixed int to float

+++ classes.py
@@ -1,6 +1,6 @@
 class WaterUsage:
 
-    def __init__(self, shower_time: int, flushes: int, sink_time: int, housing: str):
+    def __init__(self, shower_time: float, flushes: float, sink_time: float, housing: str):
         self.shower_time = shower_time
         self.flushes = flushes
         self.sink_time = sink_time
