+++ main.py
@@ -44,7 +44,7 @@ def main():
     user_dorm_data = ""
     while user_dorm_data not in valid_dorms:
         user_dorm_data = input("What housing do you live in? (PCV, Red Bricks, North Mountain, Yakitutu, Cerro Vista, Sierra Madre): ").upper()
-    user_data = WaterUsage(int(user_shower_data), int(user_flush_data), int(user_sink_data), user_dorm_data.upper())
+    user_data = WaterUsage(float(user_shower_data), float(user_flush_data), float(user_sink_data), user_dorm_data.upper())
 
     print("\nYour average daily water usage:", functions.user_average_water_usage(user_data), "Gallons")
     for dorm in Housing_average_dictionary:
